the ultimate
If you type in hint you get a hint, if you type in vowels all of the vowels will appear, if the word is less the five characters long you will have to restart, if the word is longer then thirteen characters you will have to restart. Hope you enjoy.